requirements.txt
python-dotenv==0.21.0
requests==2.28.1
redis==4.3.5
python-telegram-bot==13.14